requirements.txt
appdirs==1.4.4
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
coverage==5.3.1
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.1
Django==3.0.5
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
flake8==3.8.4
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
isort==5.6.4
gunicorn==20.0.4
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.790
mypy-extensions==0.4.3
packaging==20.3
pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.13.0
python-telegram-bot==12.7
pytils==0.3
pytz==2020.1
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
toml==0.10.2
tornado==6.0.4
twilio==6.51.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.2.2
wcwidth==0.1.9
zipp==3.1.0
